requirements.txt
Flask==0.11.1
Flask-SocketIO==2.8.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.11
click==6.6
eventlet==0.19.0
facebook-sdk==2.0.0
greenlet==0.4.10
itsdangerous==0.24
python-engineio==1.1.0
python-socketio==1.6.1
requests==2.12.3
six==1.10.0
celery==4.0.1
redis==2.10.5
pytest==3.0.5
pytest-flask==0.10.0
textblob==0.11.1